Revert "Merge branch 'main' into rgc-integracion-escenas"
This reverts commit cc44ec953debf4a0674e0365d31ed22b95511290, reversing
changes made to 3fb0050b8f2897a73c21f7797afe3d969d9446ac.

+++ Assets/Scripts/Persistence.cs
@@ -6,9 +6,9 @@ using static UnityEngine.Rendering.DebugUI;
 
 public class Persistence
 {
-    public float hp;
-    public float ron;
-    public float internalDamage;
+    public int hp;
+    public int ron;
+    public int internalDamage;
     public int selectedConsumable;
     public int addedDamage;
     public float defense;
@@ -18,6 +18,11 @@ public class Persistence
     public bool canShoot;
     public int money;
 
+    /*
+     * 1- Es necesario añadir la persistencia de enemigos muertos y los objetos destruidos hasta el reinicio en un punto de descanso
+     * 2- Es necesario preservar aquellos mecanismos o eventos persistentes del mapa
+     * 3- Es necesario almacenar los objetos ya recolectados
+     */
     public Persistence(int hp, int ron, int internalDamage, int selectedConsumable, int addedDamage, float defense, bool hasHook, bool hasParrot, bool hasGun, bool canShoot, int money)
     {
         this.hp = hp;
@@ -41,9 +46,9 @@ public class Persistence
             return null;
         }
 
-        float hp = PlayerPrefs.GetFloat("hp");
-        float ron = PlayerPrefs.GetFloat("ron");
-        float internalDamage = PlayerPrefs.GetFloat("internalDamage");
+        int hp = PlayerPrefs.GetInt("hp");
+        int ron = PlayerPrefs.GetInt("ron");
+        int internalDamage = PlayerPrefs.GetInt("internalDamage");
         int selectedConsumable = PlayerPrefs.GetInt("selectedConsumable");
         int addedDamage = PlayerPrefs.GetInt("addedDamage");
         float defense = PlayerPrefs.GetFloat("defense");
@@ -60,9 +65,9 @@ public class Persistence
     }
     public void SavePersistence() 
     {
-        PlayerPrefs.SetFloat("hp", hp);
-        PlayerPrefs.SetFloat("ron", ron);
-        PlayerPrefs.SetFloat("internalDamage", internalDamage);
+        PlayerPrefs.SetInt("hp", hp);
+        PlayerPrefs.SetInt("ron", ron);
+        PlayerPrefs.SetInt("internalDamage", internalDamage);
         PlayerPrefs.SetInt("selectedConsumable", selectedConsumable);
         PlayerPrefs.SetInt("addedDamage", addedDamage);
         PlayerPrefs.SetFloat("defense", defense);
